split Device Annoucement into a dedicated module
